Abstract template logic into address model method
Business logic doesn't belong in templates.  This method is easier to
read and more easily reusable (see [Don't Repeat Yourself][1]).

See documentation on built-in [linebreaksbr][2] template filter.  For
information on why the address method is called when used as a template
variable see [rendering a context][3] in the template documentation.

[1]: http://c2.com/cgi/wiki?DontRepeatYourself
[2]: https://docs.djangoproject.com/en/1.4/ref/templates/builtins/#std:templatefilter-linebreaksbr
[3]: https://docs.djangoproject.com/en/1.4/ref/templates/api/#rendering-a-context

Demonstrates:
* First filter (linebreaksbr)
* Principle of keeping tricky logic out of templates